Add options [connections].connection_timeout, [connections].ftp_response_timeout, [connections].timeout.
Set the number of seconds to wait while trying to connect. Use 0 to wait
indefinitely. Pass a long. It should contain the maximum time in seconds that
you allow the connection to the server to take. This only limits the connection
phase, once it has connected, this option is of no more use. Set to zero to
disable connection timeout (it will then only timeout on the system's internal
timeouts). See also the TIMEOUT option.
default:
connection_timeout=15

Set a timeout period (in seconds) on the amount of time that the server is
allowed to take in order to generate a response message for a command before the
session is considered hung. While awaiting for a response, this value overrides
TIMEOUT. It is recommended that if used in conjunction with TIMEOUT, you set
FTP_RESPONSE_TIMEOUT to a value smaller than TIMEOUT.
default:
ftp_response_timeout=5000

maximum amount of time to download segment in seconds
Set the maximum number of seconds for a connection to execute.
Pass a long as parameter containing the maximum time in seconds that you allow
the transfer operation to take. Normally, name lookups can take a considerable
time and limiting operations to less than a few minutes risk aborting perfectly
normal operations.
default:
timeout=6000